course/app/pyproject.toml
[project]
name = "faq-agent-app"
version = "0.1.0"
description = "Day 6: Streamlit + CLI for the DataTalks FAQ agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minsearch>=0.0.10",
    "openai>=2.30.0",
    "pydantic-ai>=1.73.0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "requests>=2.33.0",
    "streamlit>=1.40.0",
]
